iss/requirements.txt
annotated-types==0.6.0
anyio==4.1.0
asn1crypto==1.5.1
cbor2==5.5.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
ndg-httpsclient==0.5.1
packaging==23.2
pusher==3.3.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
python-multipart==0.0.17
redis==5.0.1
requests==2.31.0
shortuuid==1.0.11
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
starlette==0.31.1
trycourier==5.0.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
webauthn==1.11.1
